+++ ventilation_company/database/models/work_catalog.py
@@ -1,8 +1,10 @@
+from __future__ import annotations
 """ORM-РјРѕРґРµР»СЊ РґР»СЏ РєР°С‚Р°Р»РѕРіСѓ СЂРѕР±С–С‚."""
 
-from __future__ import annotations
+from decimal import Decimal
 
-from sqlalchemy import Float, Index, Integer, String
+
+from sqlalchemy import Float, Index, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ventilation_company.database.base import Base
@@ -13,7 +15,7 @@ class WorkCatalog(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String, unique=True, nullable=False)
-    price: Mapped[float] = mapped_column(Float, nullable=False)
+    price: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     unit: Mapped[str] = mapped_column(String, nullable=False)
     description: Mapped[str | None] = mapped_column(String, nullable=True)
     is_active: Mapped[int] = mapped_column(Integer, default=1)
